# API2_ZOO/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API que proporciona datos en tiempo real sobre visitas y atracciones, protegida con autenticación por token JWT PARA PODER USAR LOS METODOS SE NECESITARA UN TOKEN EL CUAL UNA VEZ OBTENIDO SE INSERTARA EN EL APARTADO DE AUTHORIZE MAS ESPECIFICO EL BOTON VERDE DE ABAJO CON UN CANDADO INSERTAR EL TOKEN EN EL APARTADO VALUE Y PULSAR EL BOTON DE AUTHORIZE UNA VEZ HECHO ESO DEBERIA DE PODER CERRAR LA VENTANA SIN PROBLEMAS Y USAR LOS METODOS DE AHORA EN ADELANTE",
        "title": "API de Monitoreo de Visitas y Atracciones",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/atraccion/lastweek": {
            "get": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "atraccion"
                ],
                "summary": "Obtener atracciones de las 6 fechas más recientes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Atraccion"
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/atraccion/now": {
            "get": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "atraccion"
                ],
                "summary": "Obtener atracciones de la fecha más reciente",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Atraccion"
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/atraccion/ojiva": {
            "get": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "atraccion"
                ],
                "summary": "Obtener ojiva de atracción (tiempo total por hora)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha en formato YYYY-MM-DD",
                        "name": "fecha",
                        "in": "query"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/atraccion/yesterday": {
            "get": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "atraccion"
                ],
                "summary": "Obtener atracciones de la penúltima fecha registrada",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Atraccion"
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/visitas/lastweek": {
            "get": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitas"
                ],
                "summary": "Obtener visitas de las 6 fechas más recientes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Visitas"
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/visitas/now": {
            "get": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitas"
                ],
                "summary": "Obtener visitas de la fecha más reciente",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Visitas"
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/visitas/ojiva": {
            "get": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitas"
                ],
                "summary": "Obtener ojiva de visitas (suma por hora)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha en formato YYYY-MM-DD",
                        "name": "fecha",
                        "in": "query"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/visitas/yesterday": {
            "get": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitas"
                ],
                "summary": "Obtener visitas de la penúltima fecha registrada",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Visitas"
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/visitasgeneral": {
            "get": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitasGeneral"
                ],
                "summary": "Obtener todas las visitas registradas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.VisitaGeneral"
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitasGeneral"
                ],
                "summary": "Crear una nueva visita",
                "parameters": [
                    {
                        "description": "Datos de la visita",
                        "name": "visita",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.VisitaGeneral"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/visitasgeneral/{fecha}": {
            "get": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitasGeneral"
                ],
                "summary": "Obtener una visita por fecha",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha de la visita en formato YYYY-MM-DD",
                        "name": "fecha",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.VisitaGeneral"
                        }
                    },
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitasGeneral"
                ],
                "summary": "Actualizar una visita por fecha",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha de la visita en formato YYYY-MM-DD",
                        "name": "fecha",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevos datos de la visita",
                        "name": "visita",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.VisitaGeneral"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Token invál",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWTQuery": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitasGeneral"
                ],
                "summary": "Eliminar una visita por fecha",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha de la visita en formato YYYY-MM-DD",
                        "name": "fecha",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o no autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.Atraccion": {
            "type": "object",
            "properties": {
                "enviado": {
                    "type": "boolean"
                },
                "fecha": {
                    "type": "string"
                },
                "hora": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "tiempo": {
                    "type": "integer"
                },
                "zona": {
                    "type": "string"
                }
            }
        },
        "entities.VisitaGeneral": {
            "type": "object",
            "properties": {
                "fecha": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "visitas": {
                    "type": "integer"
                }
            }
        },
        "entities.Visitas": {
            "type": "object",
            "properties": {
                "enviado": {
                    "type": "boolean"
                },
                "fecha": {
                    "type": "string"
                },
                "hora": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "visitantes": {
                    "type": "integer"
                },
                "zona": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWTQuery": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}